move channel tests to modules
